+++ Apermo/Tests/WordPress/SwitchToBlogRequiresRestoreUnitTest.php
@@ -1,0 +1,42 @@
+<?php
+declare(strict_types=1);
+
+/**
+ * Unit test for the SwitchToBlogRequiresRestore sniff.
+ *
+ * @package Apermo\Tests\WordPress
+ */
+
+namespace Apermo\Tests\WordPress;
+
+use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
+
+/**
+ * Unit test for Apermo.WordPress.SwitchToBlogRequiresRestore.
+ */
+class SwitchToBlogRequiresRestoreUnitTest extends AbstractSniffUnitTest {
+
+	/**
+	 * Returns the lines where errors should occur.
+	 *
+	 * @param string $testFile The name of the test file being tested.
+	 *
+	 * @return array<int, int>
+	 */
+	protected function getErrorList( $testFile = '' ) {
+		return [
+			7 => 1, // switch_to_blog without restore.
+		];
+	}
+
+	/**
+	 * Returns the lines where warnings should occur.
+	 *
+	 * @param string $testFile The name of the test file being tested.
+	 *
+	 * @return array<int, int>
+	 */
+	protected function getWarningList( $testFile = '' ) {
+		return [];
+	}
+}